+++ source/Classes/Block/32_Toggle.ps1
@@ -5,6 +5,8 @@ class Toggle_structure
     [notion_block[]] $children = @()
     #TODO: Implement addchildren
 
+    Toggle_structure() {}
+
     Toggle_structure([rich_text[]] $rich_text, [notion_color] $color = "default")
     {
         $this.rich_text = @($rich_text)
@@ -22,7 +24,12 @@ class notion_toggle_block : notion_block
     [notion_blocktype] $type = "toggle"
     [Toggle_structure] $toggle
 
-    notion_toggle_block([rich_text[]] $rich_text, [notion_color] $color = "default")
+    notion_toggle_block() : base("toggle")
+    {
+
+    }
+
+    notion_toggle_block([rich_text[]] $rich_text, [notion_color] $color = "default") : base("toggle")
     {
         $this.toggle = [Toggle_structure]::new($rich_text, $color)
     }